requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
pydantic-core==2.23.4
SQLAlchemy==2.0.31
asyncpg==0.29.0
psycopg[binary]==3.2.1
alembic==1.13.2
python-dotenv==1.0.1
redis>=5.0